backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
python-dotenv==0.19.0
dashscope==1.10.0
oss2==2.18.1
pymysql==1.0.2
SQLAlchemy==1.4.23
soundfile==0.12.1
numpy
requests==2.26.0
websockets==10.0
ffmpeg-python==0.2.0
wave==0.0.2 